+++ src/main/java/ru/skypro/homework/service/ImageService.java
@@ -5,9 +5,12 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 import ru.skypro.homework.entity.Ads;
 import ru.skypro.homework.entity.Image;
+import ru.skypro.homework.entity.User;
 import ru.skypro.homework.exception.AdsNotFoundException;
+import ru.skypro.homework.exception.RequestDeniedException;
 import ru.skypro.homework.repository.AdsRepository;
 import ru.skypro.homework.repository.ImageRepository;
+import ru.skypro.homework.repository.UserRepository;
 
 import java.io.IOException;
 import java.nio.file.Files;
@@ -21,13 +24,15 @@ public class ImageService {
     private final String imageDir;
     private final ImageRepository imageRepository;
     private final AdsRepository adsRepository;
+    private final UserRepository userRepository;
     private final DateTimeFormatter format = DateTimeFormatter.ofPattern("yyyy-MM-dd");
 
     public ImageService(@Value("${path.to.images.folder}.") String imageDir, ImageRepository imageRepository,
-                        AdsRepository adsRepository) {
+                        AdsRepository adsRepository, UserRepository userRepository) {
         this.imageDir = imageDir;
         this.imageRepository = imageRepository;
         this.adsRepository = adsRepository;
+        this.userRepository = userRepository;
     }
 
     public Image addImage(Ads ads, MultipartFile multipartFile) {
@@ -44,9 +49,16 @@ public class ImageService {
             throw new RuntimeException(e);
         }
     }
+    public byte[] getImage(int idAds) {
+        Ads ads = adsRepository.findById(idAds).orElseThrow(AdsNotFoundException::new);
+        Image image = ads.getImage();
+        return image.getBytes();
 
-    public byte[] updateImage(int id, MultipartFile image) {
-        Ads ads = adsRepository.findById(id).orElseThrow(AdsNotFoundException::new);
+    }
+    public byte[] updateImage(String username, int idAds, MultipartFile image) {
+        User user = userRepository.findByUserName(username);
+        Ads ads = adsRepository.findById(idAds).orElseThrow(AdsNotFoundException::new);
+        if(!ads.getUser().equals(user)) {throw new RequestDeniedException();}
         Image oldImage = ads.getImage();
         imageRepository.delete(oldImage);
         addImage(ads, image);
